+++ ng-env.d.ts
@@ -1,0 +1,3 @@
+export declare let NG_ENV: {
+    [name: string]: string;
+};