+++ ng-env.js
@@ -1,0 +1,13 @@
+/**
+ * @fileoverview added by tsickle
+ * @suppress {checkTypes,extraRequire,missingOverride,missingReturn,unusedPrivateMembers,uselessCode} checked by tsc
+ */
+/** @type {?} */
+let globalVariable = {};
+try {
+    globalVariable = (/** @type {?} */ (window));
+}
+catch (e) { }
+/** @type {?} */
+export let NG_ENV = globalVariable.NG_ENV;
+//# sourceMappingURL=ng-env.js.map