+++ src/common/deprecated.ts
@@ -1,5 +1,12 @@
 import { deprecate } from 'util';
 
+/**
+ * Deprecate a method.
+ * 
+ * @param message - The deprecation message.
+ * 
+ * @public
+ */
 export function deprecated(message: string): MethodDecorator {
   return (target: any, propertyKey: string | symbol, descriptor: PropertyDescriptor) => {
     // tslint:disable-next-line