+++ src/process-env-configuration.ts
@@ -2,6 +2,8 @@ import { Configuration } from './configuration';
 
 /**
  * Discover and apply configuration via environment variables discovered via process.env usage.
+ * 
+ * @public
  */
 export class ProcessEnvConfiguration extends Configuration {
   protected discoverVariables(fileContent: string): string[] {